setup network fixtures moved to specific file, wait genesis to utils
