+++ tests/block_atx/add_node/test_blocks_add_node.py
@@ -2,7 +2,7 @@ from pytest_testconfig import config as testconfig
 
 from tests import queries as q
 from tests.setup_utils import add_multi_clients
-from tests.test_bs import setup_network, wait_genesis
+from tests.setup_network import setup_network
 from tests.utils import validate_blocks_per_nodes, get_pod_id, get_conf
