+++ tests/hare/test_hare.py
@@ -1,16 +1,18 @@
+from datetime import datetime, timedelta
 import time
 import pytest
+import pytz
 from pytest_testconfig import config as testconfig
 
 from tests import deployment
 from tests.conftest import DeploymentInfo
 from tests.hare.assert_hare import expect_hare, assert_all, get_max_mem_usage
 from tests.misc import CoreV1ApiClient
-from tests.test_bs import wait_genesis
 from tests.setup_utils import setup_bootstrap_in_namespace, setup_clients_in_namespace
-from tests.utils import get_curr_ind
+from tests.utils import get_curr_ind, wait_genesis
 
 ORACLE_DEPLOYMENT_FILE = './k8s/oracle.yml'
+GENESIS_TIME = pytz.utc.localize(datetime.utcnow() + timedelta(seconds=testconfig['genesis_delta']))
 
 
 def setup_server(deployment_name, deployment_file, namespace):
@@ -66,6 +68,7 @@ def setup_bootstrap_for_hare(request, init_session, setup_oracle):
 @pytest.fixture(scope='module')
 def setup_clients_for_hare(request, init_session, setup_oracle, setup_bootstrap_for_hare):
     client_info = DeploymentInfo(dep_id=setup_bootstrap_for_hare.deployment_id)
+    wait_genesis(GENESIS_TIME, testconfig['genesis_delta'])
     return setup_clients_in_namespace(testconfig['namespace'],
                                       setup_bootstrap_for_hare.pods[0],
                                       client_info,
@@ -84,7 +87,7 @@ EFK_LOG_PROPAGATION_DELAY = 20
 
 
 # simple sanity test run for one layer
-def test_hare_sanity(init_session, setup_bootstrap_for_hare, setup_clients_for_hare, wait_genesis):
+def test_hare_sanity(init_session, setup_bootstrap_for_hare, setup_clients_for_hare):
     current_index = get_curr_ind()
     # Need to wait for 1 full iteration + the time it takes the logs to propagate to ES
     round_duration = int(testconfig['client']['args']['hare-round-duration-sec'])
@@ -103,7 +106,7 @@ EXPECTED_MAX_MEM = 300*1024*1024  # MB
 
 
 # scale test run for 3 layers
-def test_hare_scale(init_session, setup_bootstrap_for_hare, setup_clients_for_hare, wait_genesis):
+def test_hare_scale(init_session, setup_bootstrap_for_hare, setup_clients_for_hare):
     current_index = get_curr_ind()
     total = int(testconfig['client']['replicas']) + int(testconfig['bootstrap']['replicas'])
