+++ tests/tx_generator/test_tx_generator.py
@@ -3,9 +3,9 @@ import pprint
 import random
 import time
 
+from tests.setup_network import setup_network
 from tests.tx_generator import actions
 from tests.tx_generator.models.accountant import Accountant
-from tests.test_bs import setup_network, wait_genesis
 import tests.tx_generator.config as conf
 from tests.tx_generator.models.wallet_api import WalletAPI
